Revert: `TestVasp` is better than `TestVaspBase`
